Move a ton of kotlin files around
