+++ appindexing/app/src/main/java/com/google/firebase/example/appindexing/kotlin/AppIndexingUpdateReceiver.kt
@@ -1,13 +1,14 @@
-package com.google.firebase.example.appindexing
+package com.google.firebase.example.appindexing.kotlin
 
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
 import com.google.firebase.appindexing.FirebaseAppIndex
+import com.google.firebase.example.appindexing.AppIndexingUpdateService
 
 // [START appindexing_update_receiver]
 /** Receives broadcast for App Indexing Update. */
-class KotlinAppIndexingUpdateReceiver : BroadcastReceiver() {
+class AppIndexingUpdateReceiver : BroadcastReceiver() {
 
     override fun onReceive(context: Context, intent: Intent?) {
         if (intent != null && FirebaseAppIndex.ACTION_UPDATE_INDEX == intent.action) {
