+++ crashlytics/app/src/main/AndroidManifest.xml
@@ -22,7 +22,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".KotlinMainActivity" />
+        <activity android:name=".kotlin.MainActivity" />
     </application>
 
 </manifest>
