+++ firebaseoptions/app/src/main/java/devrel/firebase/google/com/firebaseoptions/kotlin/MainActivity.kt
@@ -1,12 +1,13 @@
-package devrel.firebase.google.com.firebaseoptions
+package devrel.firebase.google.com.firebaseoptions.kotlin
 
 import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
 import com.google.firebase.FirebaseApp
 import com.google.firebase.FirebaseOptions
 import com.google.firebase.database.FirebaseDatabase
+import devrel.firebase.google.com.firebaseoptions.R
 
-class KotlinMainActivity : AppCompatActivity() {
+class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
