+++ perf/app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name="com.google.firebase.example.perf.KotlinMainActivity" />
+        <activity android:name="com.google.firebase.example.perf.kotlin.MainActivity" />
     </application>
 
 </manifest>
