+++ perf/app/src/main/java/com/google/firebase/example/perf/MainActivity.java
@@ -7,9 +7,9 @@ import android.support.v7.app.AppCompatActivity;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.example.perf.interfaces.MainActivityInterface;
-import com.google.firebase.example.perf.model.Item;
-import com.google.firebase.example.perf.model.ItemCache;
-import com.google.firebase.example.perf.model.User;
+import com.google.firebase.example.perf.kotlin.model.Item;
+import com.google.firebase.example.perf.kotlin.model.ItemCache;
+import com.google.firebase.example.perf.kotlin.model.User;
 import com.google.firebase.perf.FirebasePerformance;
 import com.google.firebase.perf.metrics.AddTrace;
 import com.google.firebase.perf.metrics.HttpMetric;
