+++ perf/app/src/main/java/com/google/firebase/example/perf/kotlin/model/Item.kt
@@ -1,0 +1,4 @@
+package com.google.firebase.example.perf.kotlin.model
+
+class Item {
+}
