--- perf/app/src/main/java/com/google/firebase/example/perf/model/Item.kt
@@ -1,4 +1,0 @@
-package com.google.firebase.example.perf.model
-
-class Item {
-}
